+++ autofill/autofill-impl/src/main/java/com/duckduckgo/app/email/service/DuckAddressStatusManagementService.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.email.api
+package com.duckduckgo.app.email.service
 
 import com.duckduckgo.anvil.annotations.ContributesNonCachingServiceApi
 import com.duckduckgo.di.scopes.AppScope
