+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/ui/credential/repository/RemoteDuckAddressStatusRepository.kt
@@ -17,7 +17,7 @@
 package com.duckduckgo.autofill.impl.ui.credential.repository
 
 import com.duckduckgo.app.email.EmailManager
-import com.duckduckgo.app.email.api.DuckAddressStatusManagementService
+import com.duckduckgo.app.email.service.DuckAddressStatusManagementService
 import com.duckduckgo.app.global.DispatcherProvider
 import com.duckduckgo.autofill.impl.ui.credential.repository.DuckAddressStatusRepository.ActivationStatusResult
 import com.duckduckgo.autofill.impl.ui.credential.repository.DuckAddressStatusRepository.ActivationStatusResult.Activated
